requirements.txt
streamlit
crewai
langchain-openai
python-dotenv
requests
crewai-tools
fpdf
tweepy
youtube-transcript-api
duckduckgo-search